Cleanup and signed/unsigned corrections (#10)
* Replace SNOWBOY_ERROR by proper exceptions
* Cleaning up lots of signed/unsigned
* More signed/unsigned work
* Enabled Wextra
* Cleanup pipeline-template-cut

+++ lib/dtw-lib.h
@@ -31,13 +31,13 @@ namespace snowboy {
 		void SetOptions(const SlidingDtwOptions&);
 		void SetEarlyStopThreshold(float);
 		void Reset();
-		int GetWindowSize() const;
+		size_t GetWindowSize() const;
 		float GetDistance(int, int) const;
 		float ComputeVectorDistance(const VectorBase&, const VectorBase&) const;
 		float ComputeDtwDistance(int, const MatrixBase&);
-		void ComputeBandBoundary(int, int*, int*) const;
+		void ComputeBandBoundary(int, size_t*, size_t*) const;
 		virtual ~SlidingDtw();
 	};
 
-	float DtwAlign(DistanceType, const MatrixBase&, const MatrixBase&, std::vector<std::vector<int>>*);
+	float DtwAlign(DistanceType, const MatrixBase&, const MatrixBase&, std::vector<std::vector<size_t>>*);
 } // namespace snowboy
