+++ lib/vad-lib.h
@@ -1,10 +1,11 @@
 #pragma once
+#include <cstdint>
 #include <vector>
 
 namespace snowboy {
 	struct VadStateOptions {
-		int min_non_voice_frames;
-		int min_voice_frames;
+		uint32_t min_non_voice_frames;
+		uint32_t min_voice_frames;
 	};
 	enum VoiceType { VT_0,
 					 VT_1,
@@ -15,7 +16,7 @@ namespace snowboy {
 	struct VadState {
 		VadStateOptions m_options;
 		bool m_field_x10;
-		int m_field_x14;
+		size_t m_field_x14;
 
 		VadState(VadStateOptions options);
 		virtual ~VadState();
